Introducting navigation options and (some) dependency injection.
