+++ WpfApp/View/MainWindow.xaml
@@ -6,46 +6,29 @@
         xmlns:vm="clr-namespace:WpfApp.ViewModel"
         xmlns:local="clr-namespace:WpfApp.View"
         mc:Ignorable="d"
-        Title="Notepad" Height="1080" Width="1920" WindowStyle="SingleBorderWindow" WindowStartupLocation="Manual">
+        Title="Notepad" Height="900" Width="1600" WindowStyle="SingleBorderWindow" WindowStartupLocation="Manual" WindowState="Maximized">
+    <Window.DataContext>
+        <vm:MainViewModel/>
+    </Window.DataContext>
     <Grid>
-        <Grid.Resources>
-            <Style TargetType="{x:Type Button}" x:Key="formatTextStyle">
-                <Setter Property="FontFamily" Value="Palatino Linotype"></Setter>
-                <Setter Property="Width" Value="30"></Setter>
-                <Setter Property="FontSize" Value ="14"></Setter>
-                <Setter Property="CommandTarget" Value="{Binding ElementName=mainRTB}"></Setter>
-            </Style>
-        </Grid.Resources>
-
-        <Grid.ColumnDefinitions>
-            <ColumnDefinition/>
-            <ColumnDefinition/>
-            <ColumnDefinition/>
-        </Grid.ColumnDefinitions>
-
         <Grid.RowDefinitions>
             <RowDefinition Height="auto"/>
-            <RowDefinition Height="auto"/>
-            <RowDefinition Height="1*"/>
-            <RowDefinition Height="auto"/>
+            <RowDefinition Height="*"/>
         </Grid.RowDefinitions>
 
-        <StackPanel Orientation="Horizontal" Grid.Column="0" Grid.Row="0" Margin="10,10,10,5">
-            <Button Width="64" Height="64" Margin="0,0,5,0" Content="B" FontSize="24" Command="{Binding BoldToggle}" Background="{Binding BoldBackground}">
-                <Button.InputBindings>
-                    <KeyBinding 
-                        Gesture="Ctrl+B"
-                        Command="{Binding BoldToggle}"/>
-                </Button.InputBindings>
-            </Button>
-            <Button Width="64" Height="64" Margin="5,0,5,0" Content="U" FontSize="24" Command="{Binding UnderlineToggle}" Background="{Binding UnderlineBackground}"/>
-            <Button Width="64" Height="64" Margin="5,0,0,0" Content="I" FontSize="24" Command="{Binding ItalicToggle}" Background="{Binding ItalicBackground}"/>
+        <Grid.Resources>
+            <DataTemplate DataType="{x:Type vm:NotepadViewModel}">
+                <local:NotepadView/>
+            </DataTemplate>
+            <DataTemplate DataType="{x:Type vm:DataVisualiserViewModel}">
+                <local:DataVisualiser/>
+            </DataTemplate>
+        </Grid.Resources>
+        <ContentControl Content="{Binding CurrentView}" Grid.Row="1" />
+
+        <StackPanel>
+            <Button Content="Home" Command="{Binding NavigateToHome}"/>
+            <Button Content="Data Visualisation" Command="{Binding NavigateToVisualiser}"/>
         </StackPanel>
-
-        <Separator Grid.Column="0" Grid.ColumnSpan="3" Grid.Row="1" Margin="5,5,5,5"/>
-
-        <RichTextBox x:Name="mainRTB" Grid.Column="0" Grid.ColumnSpan="3" Grid.Row="2" Margin="10,5,10,5" FontSize="16"/>
-
-        <Button Grid.Column="3" Grid.Row="3" Height="64" Margin="10,5,10,10" Content="Save" FontSize="24"/>
     </Grid>
 </Window>
