+++ WpfApp/View/NotepadView.xaml
@@ -1,0 +1,51 @@
+﻿<UserControl x:Class="WpfApp.View.NotepadView"
+             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
+             xmlns:vm="clr-namespace:WpfApp.ViewModel"
+             xmlns:local="clr-namespace:WpfApp.View"
+             mc:Ignorable="d" 
+             d:DesignHeight="450" d:DesignWidth="800">
+    <Grid>
+        <Grid.Resources>
+            <Style TargetType="{x:Type Button}" x:Key="formatTextStyle">
+                <Setter Property="FontFamily" Value="Palatino Linotype"></Setter>
+                <Setter Property="Width" Value="30"></Setter>
+                <Setter Property="FontSize" Value ="14"></Setter>
+                <Setter Property="CommandTarget" Value="{Binding ElementName=mainRTB}"></Setter>
+            </Style>
+        </Grid.Resources>
+
+        <Grid.ColumnDefinitions>
+            <ColumnDefinition/>
+            <ColumnDefinition/>
+            <ColumnDefinition/>
+        </Grid.ColumnDefinitions>
+
+        <Grid.RowDefinitions>
+            <RowDefinition Height="auto"/>
+            <RowDefinition Height="auto"/>
+            <RowDefinition Height="1*"/>
+            <RowDefinition Height="auto"/>
+        </Grid.RowDefinitions>
+
+        <StackPanel Orientation="Horizontal" Grid.Column="0" Grid.Row="0" Margin="10,10,10,5">
+            <Button Width="64" Height="64" Margin="0,0,5,0" Content="B" FontSize="24" Command="{Binding BoldToggle}" Background="{Binding BoldBackground}">
+                <Button.InputBindings>
+                    <KeyBinding 
+                        Gesture="Ctrl+B"
+                        Command="{Binding BoldToggle}"/>
+                </Button.InputBindings>
+            </Button>
+            <Button Width="64" Height="64" Margin="5,0,5,0" Content="U" FontSize="24" Command="{Binding UnderlineToggle}" Background="{Binding UnderlineBackground}"/>
+            <Button Width="64" Height="64" Margin="5,0,0,0" Content="I" FontSize="24" Command="{Binding ItalicToggle}" Background="{Binding ItalicBackground}"/>
+        </StackPanel>
+
+        <Separator Grid.Column="0" Grid.ColumnSpan="3" Grid.Row="1" Margin="5,5,5,5"/>
+
+        <RichTextBox Grid.Column="0" Grid.ColumnSpan="3" Grid.Row="2" Margin="10,5,10,5" FontSize="16"/>
+
+        <Button Grid.Column="3" Grid.Row="3" Height="64" Margin="10,5,10,10" Content="Save" FontSize="24"/>
+    </Grid>
+</UserControl>
