+++ WpfApp/View/NotepadView.xaml.cs
@@ -1,8 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Linq;
 using System.Text;
-using System.Threading.Tasks;
 using System.Windows;
 using System.Windows.Controls;
 using System.Windows.Data;
@@ -16,11 +14,11 @@ using System.Windows.Shapes;
 namespace WpfApp.View
 {
     /// <summary>
-    /// Interaction logic for MainWindow.xaml
+    /// Interaction logic for NotepadView.xaml
     /// </summary>
-    public partial class MainWindow : Window
+    public partial class NotepadView : UserControl
     {
-        public MainWindow()
+        public NotepadView()
         {
             InitializeComponent();
         }
