+++ WpfApp/ViewModel/DataVisualiserViewModel.cs
@@ -1,0 +1,21 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+using WpfApp.ViewModel.Helper;
+using WpfApp.ViewModel.Interfaces;
+
+namespace WpfApp.ViewModel
+{
+    public class DataVisualiserViewModel : OnPropertyChangedImplementation, INavigatable
+    {
+        public void OnNavigateAway()
+        {
+            throw new NotImplementedException();
+        }
+
+        public void OnNavigateTo()
+        {
+            throw new NotImplementedException();
+        }
+    }
+}
