+++ WpfApp/ViewModel/Interfaces/INavigatable.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace WpfApp.ViewModel.Interfaces
+{
+    public interface INavigatable
+    {
+        void OnNavigateTo();
+
+        void OnNavigateAway();
+    }
+}
